Unit test

+++ src/main/java/com/cash4books/cash4books/dto/cart/UserCartDto.java
@@ -3,14 +3,14 @@ package com.cash4books.cash4books.dto.cart;
 import java.util.List;
 
 public class UserCartDto {
-    private int total;
+    private double total;
     private List<CartDto> cartDetails;
 
-    public int getTotal() {
+    public double getTotal() {
         return total;
     }
 
-    public void setTotal(int total) {
+    public void setTotal(double total) {
         this.total = total;
     }
